+++ tasks.rb
@@ -166,6 +166,7 @@ def create_hpxmls
     'base-enclosure-windows-none.xml' => 'base.xml',
     'base-enclosure-windows-physical-properties.xml' => 'base.xml',
     'base-enclosure-windows-shading.xml' => 'base.xml',
+    'base-enclosure-thermal-mass.xml' => 'base.xml',
     'base-foundation-ambient.xml' => 'base.xml',
     'base-foundation-basement-garage.xml' => 'base.xml',
     'base-foundation-complex.xml' => 'base.xml',
@@ -2868,6 +2869,12 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml.windows[3].exterior_shading_factor_winter = 1.0
     hpxml.windows[3].interior_shading_factor_summer = 0.0
     hpxml.windows[3].interior_shading_factor_winter = 1.0
+  elsif ['base-enclosure-thermal-mass.xml'].include? hpxml_file
+    hpxml.partition_wall_mass.area_fraction = 0.8
+    hpxml.partition_wall_mass.interior_finish_type = HPXML::InteriorFinishGypsumBoard
+    hpxml.partition_wall_mass.interior_finish_thickness = 0.25
+    hpxml.furniture_mass.area_fraction = 0.8
+    hpxml.furniture_mass.type = HPXML::FurnitureMassTypeHeavyWeight
   elsif ['base-misc-defaults.xml'].include? hpxml_file
     hpxml.attics.reverse_each do |attic|
       attic.delete